refactor: move ModuleContentEditor component back to course module page
